requirements.txt
streamlit>=1.30.0
groq>=0.4.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
python-docx>=1.1.0
openpyxl>=3.1.0
reportlab>=4.0.0
pandas>=2.0.0
